Changes
- Made rarity a double. The String version is taken from the double as an integer (rounded down). The double allows for better sorting between titles of the same rarity
- Made cache system for reverse and rarity, so the values persist over pages
- Added sounds to some GUI buttons
- Fixed titles with hex colors not being sorted by alphabetical order correctly

+++ src/main/java/com/strangeone101/playertitles/Config.java
@@ -27,7 +27,7 @@ public class Config {
         for (String titleId : titleConfig.getKeys(false)) {
             String name = PlayerTitlesPlugin.color(titleConfig.getString(titleId + ".name", "????"));
             String desc = PlayerTitlesPlugin.color(titleConfig.getString(titleId + ".description", ""));
-            int rarity = titleConfig.getInt(titleId + ".rarity", 1);
+            float rarity = (float) titleConfig.getDouble(titleId + ".rarity", 1);
             String group = titleConfig.getString(titleId + ".group");
 
             Title title;
